Extract pgflow-core specific typescript code from pkgs/edge-worker to pkgs/core
