+++ pkgs/core/package.json
@@ -17,5 +17,8 @@
   "module": "./dist/index.js",
   "devDependencies": {
     "supabase": "^2.6.8"
+  },
+  "dependencies": {
+    "postgres": "^3.4.5"
   }
 }
